+++ src/index.js
@@ -1,0 +1,5 @@
+import RecaptchaWrapper from "./recaptcha-wrapper";
+import ReCAPTCHA from "./recaptcha";
+
+export default RecaptchaWrapper;
+export { ReCAPTCHA };