+++ src/recaptcha-wrapper.js
@@ -7,9 +7,7 @@ function getOptions() {
 function getURL() {
   const dynamicOptions = getOptions();
   const lang = dynamicOptions.lang ? `&hl=${dynamicOptions.lang}` : "";
-  const hostname = dynamicOptions.useRecaptchaNet
-    ? "recaptcha.net"
-    : "www.google.com";
+  const hostname = dynamicOptions.useRecaptchaNet ? "recaptcha.net" : "www.google.com";
   return `https://${hostname}/recaptcha/api.js?onload=${callbackName}&render=explicit${lang}`;
 }
 
@@ -17,9 +15,8 @@ const callbackName = "onloadcallback";
 const globalName = "grecaptcha";
 const initialOptions = getOptions();
 
-export default makeAsyncScriptLoader(ReCAPTCHA, getURL, {
+export default makeAsyncScriptLoader(getURL, {
   callbackName,
   globalName,
-  removeOnMount: initialOptions.removeOnMount || false,
-  exposeFuncs: ["getValue", "getWidgetId", "reset", "execute"],
-});
+  removeOnUnmount: initialOptions.removeOnUnmount || false,
+})(ReCAPTCHA);